pyproject.toml
[tool.poetry]
name = "fast-api-ml"
version = "0.1.0"
description = ""
authors = ["Jung Bak <jung-bak@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.3.0"
fastapi = "^0.110.1"
torch = "^2.2.2"
transformers = "^4.39.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
